+++ tests/unit/test_main.py
@@ -1,5 +1,6 @@
 import md2cf.__main__ as main
 from md2cf.api import MinimalConfluence as Confluence
+from md2cf.document import Page
 
 
 def test_upsert_page(mocker):
@@ -10,30 +11,31 @@ def test_upsert_page(mocker):
     parent_page_mock.id = mocker.sentinel.parent_page_id
     confluence.get_page.side_effect = [None, parent_page_mock]
 
-    space = mocker.sentinel.space
-    title = mocker.sentinel.title
-    body = mocker.sentinel.body
-    parent = mocker.sentinel.parent
+    page = Page(
+        space=mocker.sentinel.space,
+        title=mocker.sentinel.title,
+        body=mocker.sentinel.body,
+        parent_title=mocker.sentinel.parent,
+    )
+
     message = mocker.sentinel.message
 
     main.upsert_page(
-        confluence=confluence,
-        space=space,
-        title=title,
-        body=body,
-        parent=parent,
-        message=message
+        confluence=confluence, page=page, message=message,
     )
 
-    confluence.get_page.assert_has_calls([
-        mocker.call(title=title, space_key=space, page_id=None),
-        mocker.call(title=parent, space_key=space)
-    ], any_order=False)
+    confluence.get_page.assert_has_calls(
+        [
+            mocker.call(title=page.title, space_key=page.space, page_id=None),
+            mocker.call(title=page.parent_title, space_key=page.space),
+        ],
+        any_order=False,
+    )
 
     confluence.create_page.assert_called_once_with(
-        space=space,
-        title=title,
-        body=body,
+        space=page.space,
+        title=page.title,
+        body=page.body,
         parent_id=mocker.sentinel.parent_page_id,
-        update_message=message
+        update_message=message,
     )