+++ CHANGELOG.md
@@ -4,6 +4,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+- Add support for image attachments
+- Add support for recursive upload of directories
+- Skip empty directories
+- Collapse empty directories
+- Collapse single-document folders
+- Support folder title customization
+- Remove dependency on BeautifulSoup by finding the title during rendering
+- Can now specify text top preface to each uploaded page
+- Can now specify a page title in the YAML front matter
+- Can now specify a prefix to add to the title of each page
+- Existing pages can be moved to a different parent
+- Add dry run option
+
 ## 0.2.2 - 2019-10-22
 - Bump depdendencies